feat: add wrapper to settings form
