+++ src/pages/UserSettings/content/formSections/Focus.section.tsx
@@ -1,6 +1,6 @@
 import { Field } from 'react-final-form'
 import { useTheme } from '@emotion/react'
-import { Button, ExternalLink } from 'oa-components'
+import { ExternalLink } from 'oa-components'
 import { getSupportedProfileTypes } from 'src/modules/profile'
 import { buttons, fields, headings } from 'src/pages/UserSettings/labels'
 import { Box, Flex, Grid, Heading, Paragraph, Text } from 'theme-ui'
@@ -11,7 +11,7 @@ import { FlexSectionContainer } from './elements'
 import type { ProfileTypeLabel } from 'src/modules/profile/types'
 
 const ProfileTypes = () => {
-  const { description, error, title } = fields.activities
+  const { description, error } = fields.activities
   const theme = useTheme()
   const profileTypes = getSupportedProfileTypes().filter(({ label }) =>
     Object.keys(theme.badges).includes(label),
@@ -31,8 +31,21 @@ const ProfileTypes = () => {
               {headings.focus}
             </Heading>
           </Flex>
+
+          <Flex sx={{ alignItems: 'baseline', marginY: 2 }}>
+            <Paragraph my={4}>
+              {description}{' '}
+              <ExternalLink
+                href={theme.profileGuidelinesURL}
+                sx={{ textDecoration: 'underline', color: 'grey' }}
+                type="button"
+              >
+                {buttons.guidelines}
+              </ExternalLink>
+            </Paragraph>
+          </Flex>
+
           <Box>
-            <Paragraph my={4}>{title}</Paragraph>
             <Grid columns={['repeat(auto-fill, minmax(125px, 1fr))']} gap={2}>
               {profileTypes.map((profile, index: number) => (
                 <Box key={index}>
@@ -52,14 +65,6 @@ const ProfileTypes = () => {
                 </Box>
               ))}
             </Grid>
-            <Flex sx={{ flexWrap: 'wrap', alignItems: 'center' }} mt={4}>
-              <Text my={2}>{description}</Text>
-              <ExternalLink href={theme.profileGuidelinesURL}>
-                <Button ml={[1, 2, 2]} variant="outline" type="button">
-                  {buttons.guidelines}
-                </Button>
-              </ExternalLink>
-            </Flex>
             {props.meta.error && <Text color={theme.colors.red}>{error}</Text>}
           </Box>
         </FlexSectionContainer>
